Determination of the price.
The price of a ticket to the museum depends on the age of the user: from 7 to 2,000 soums, from 7 to 18 to 3,000 soums, from 18 to 65 to 10,000 soums, from 65 to free. Write a while loop so that the program asks for the user's age and displays the ticket price. Let the program stop when the user writes exit or quit (check both conditions).