Object-Oriented Programming. Classes and Objects.
Created new class called Auto. Added to it a few features (model, color, gearbox, price) that are relevant to cars. Gave some properties a default value (for example, mileage = 0)
Wrote methods that return the properties of an auto-related object
The get_info () method returns full information about the auto in text form.
Wrote methods that update the properties of an auto-related object.
The update_km () method takes the number and updates the mileage of the car.
Created new class called Car Showroom and filled it with the required features (showroom name, address, cars for sale).
Wrote a method to add new cars to the showroom.
Wrote a method that returns information about the cars in the showroom.
Examined the above objects and their associated techniques.