INHERITANCE AND POLYMORPHISM.
Another feature, Subject, has been added to the student class. This property was not passed as a parameter and consisted of an empty list when the object was created (self.subject = [])
A new class called Subject was created, and from this class separate objects were created for different subjects.
A new method called registration () was written for the student class. This method took the objects belonging to the Subject class as a parameter and added them to the student's list of subjects.
The remove_Subject () method was written to remove a subject from the student's subject list. This method returns a "You are not subscribed to this Subject" message if a fan not listed is passed.
In addition to the above Person class, various successor classes have been created.