Working With Files
I downloaded the file pi_million_digits.txt (in the file the number pi was written with an accuracy of one million numbers after the dot).
A function was written to determine if the date of birth is included in the number pi. For example, if the date of birth is February 25, 2000, find out if the sequence 25022000 occurs in the above text.